requirements.txt
Django==4.2.1
python-decouple==3.8
psycopg2-binary>=2.8
djangorestframework==3.14.0
django-rest-knox==4.2.0
django-model-utils==4.3.1
drf-yasg==1.21.5
gunicorn==20.1.0